Fix QAC quick typing bug
When typing quickly, it could happen that the order of responses
returned by the server was not the same as the order of requests. In
that case, the completion predictions were shown for a wrong (=previous)
question prefix.
This commit fixes this by adding a timestamp to each request and
response and discarding responses if a more recent one has already been
received.