Fix last entity vs last normal word span edit bug
Before, when the last word was an entity, removing the trailing
whitespace and adding it again resulted in the entity span being
replaced by a normal word span. The reason was, that trailing empty
normal word spans were deleted such that the whitespace was added within
the entity span. This is equivalent to changing the entity's name which
results in removal of the entity marking.
Now, an empty normal word span is not deleted if it is the last span in
the div.
The problem persists in Chrome most likely due to Chrome's inherent
contenteditable bugs.