Update + add scripts to generate wiki info mapping
Before, only information of entities which had a link to an image on
Wikidata via the P18 property was included in the wiki info mapping.
Now, information about all entities that can be mapped to one of (title,
abstract, image) is included.
Images are now either retrieved via the Wikipedia API or via a Wikidata
image property (P18, P109, P15, ...).
This commit adds the necessary scripts to create the mapping from
scratch and adds documentation about the process.
Relates to #1.